RF: Move fx, not a general helper (anymore), only used in one place
