+++ datalad/distribution/utils.py
@@ -439,17 +439,3 @@ def _handle_possible_annex_dataset(dataset, reckless):
         repo = AnnexRepo(dataset.path, init=True)
         if reckless:
             repo._run_annex_command('untrust', annex_options=['here'])
-
-
-def _save_installed_datasets(ds, installed_datasets):
-    paths = [relpath(subds.path, ds.path) for subds in installed_datasets]
-    paths_str = ", ".join(paths)
-    msg = "installed subdataset{}: {}".format(
-        "s" if len(paths) > 1 else "", paths_str)
-    lgr.info("Saving possible changes to {0} - {1}".format(
-        ds, msg))
-    ds.save(
-        files=paths + ['.gitmodules'],
-        message='[DATALAD] ' + msg,
-        all_changes=False,
-        recursive=False)
